Add system code to labels (#10)
This required reformatting the service discovery output, as each combinatorial of labels needs to be a separate array value.

It's worth noting that we need an extra array of labels values in order to have stable iteration over our map intermediary. Range over a maps keys is [non-stable in golang](https://blog.golang.org/go-maps-in-action#TOC_7.).

This is branched off https://github.com/Financial-Times/prometheus-biz-ops-service-discovery/pull/8 to avoid having to re-implement tests.